pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-migrate>=4.1.0",
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "numpy>=2.2.3",
    "opencv-python>=4.11.0.86",
    "reportlab>=4.3.1",
    "werkzeug>=3.1.3",
    "wtforms>=3.2.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.38",
]
